Add io::read_json

+++ rust/src/da.rs
@@ -8,7 +8,7 @@ use self::ndarray_odeint::*;
 pub type V = Array<f64, Ix>;
 pub type Ensemble = Vec<V>;
 
-fn teo(dt: f64, step: u32, mut x: V) -> V {
+fn teo(dt: f64, step: usize, mut x: V) -> V {
     let l = |y| lorenz63(10., 28., 8.0 / 3.0, y);
     let u = |y| rk4(&l, dt, y);
     for _ in 0..step {
@@ -17,6 +17,6 @@ fn teo(dt: f64, step: u32, mut x: V) -> V {
     x
 }
 
-pub fn forcast(xs: Ensemble, dt: f64, step: u32) -> Ensemble {
+pub fn forcast(xs: Ensemble, dt: f64, step: usize) -> Ensemble {
     xs.into_iter().map(|y| teo(dt, step, y)).collect()
 }
